move ProjectList to containers and rename Home to ProjectsView
